1260. Shift 2D Grid
